+++ packages/article-in-depth/src/article-in-depth.web.js
@@ -78,7 +78,8 @@ class ArticlePage extends Component {
       additionalRelatedArticlesFlag,
       algoliaSearchKeys,
       latestFromSectionFlag,
-      latestFromSection
+      latestFromSection,
+      olympicsKeys
     } = this.props;
 
     if (error || isLoading) {
@@ -99,6 +100,7 @@ class ArticlePage extends Component {
         algoliaSearchKeys={algoliaSearchKeys}
         latestFromSectionFlag={latestFromSectionFlag}
         latestFromSection={latestFromSection}
+        olympicsKeys={olympicsKeys}
       />
     );
   }