+++ packages/article-magazine-standard/fixtures/full-article.js
@@ -2113,11 +2113,30 @@ export const testFixture = {
     {
       byline: [
         {
+          attributes: {
+            slug: "camilla-long"
+          },
+          children: [
+            {
+              attributes: {
+                value: "Camilla Long"
+              },
+              children: [],
+              name: "text"
+            }
+          ],
+          name: "author"
+        }
+      ]
+    },
+    {
+      byline: [
+        {
           attributes: {},
           children: [
             {
               attributes: {
-                value: "Some byline"
+                value: ", Environment Editor"
               },
               children: [],
               name: "text"