+++ packages/article-main-standard/src/article-main-standard.web.js
@@ -86,7 +86,8 @@ class ArticlePage extends Component {
       additionalRelatedArticlesFlag,
       algoliaSearchKeys,
       latestFromSectionFlag,
-      latestFromSection
+      latestFromSection,
+      olympicsKeys
     } = this.props;
 
     if (error || isLoading) {
@@ -108,6 +109,7 @@ class ArticlePage extends Component {
           algoliaSearchKeys={algoliaSearchKeys}
           latestFromSectionFlag={latestFromSectionFlag}
           latestFromSection={latestFromSection}
+          olympicsKeys={olympicsKeys}
         />
       </ArticleMainStandardContainer>
     );