+++ packages/ts-components/src/components/olympics/types.ts
@@ -1,0 +1,5 @@
+export type OlympicsKeys = {
+  endpoint: string;
+  authToken: string;
+  gamesCode: string;
+};