+++ packages/depend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.27](https://github.com/newsuk/times-components/compare/@times-components/depend@2.1.26...@times-components/depend@2.1.27) (2021-07-21)
+
+**Note:** Version bump only for package @times-components/depend
+
+
+
+
+
 ## [2.1.26](https://github.com/newsuk/times-components/compare/@times-components/depend@2.1.25...@times-components/depend@2.1.26) (2019-10-11)
 
 **Note:** Version bump only for package @times-components/depend