+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.36](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.35...@times-components/pages@2.3.36) (2021-07-26)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.35](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.34...@times-components/pages@2.3.35) (2021-07-26)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.34](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.33...@times-components/pages@2.3.34) (2021-07-26)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.33](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.32...@times-components/pages@2.3.33) (2021-07-23)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.32](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.31...@times-components/pages@2.3.32) (2021-07-22)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.31](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.30...@times-components/pages@2.3.31) (2021-07-21)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.30](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.29...@times-components/pages@2.3.30) (2021-07-20)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.29](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.28...@times-components/pages@2.3.29) (2021-07-20)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 ## [2.3.28](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.27...@times-components/pages@2.3.28) (2021-07-20)
 
 **Note:** Version bump only for package @times-components/pages