+++ packages/provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.25.6](https://github.com/newsuk/times-components/compare/@times-components/provider@1.25.5...@times-components/provider@1.25.6) (2021-07-26)
+
+**Note:** Version bump only for package @times-components/provider
+
+
+
+
+
 ## [1.25.5](https://github.com/newsuk/times-components/compare/@times-components/provider@1.25.4...@times-components/provider@1.25.5) (2021-07-20)
 
 **Note:** Version bump only for package @times-components/provider